Moved tests for peak comments to separate file
