+++ tests/test_spectrum.py
@@ -1,7 +1,6 @@
 import numpy
 from matplotlib import pyplot as plt
 from matchms import Spectrum
-from matchms import Spikes
 
 
 def _assert_plots_ok(fig, n_plots):
@@ -192,17 +191,3 @@ def test_spectrum_plot_with_histogram_unspecified():
     spectrum = _create_test_spectrum()
     fig = spectrum.plot()
     _assert_plots_ok(fig, n_plots=1)
-
-
-def test_reiterating_peak_comments():
-    mz = numpy.array([100.0003, 100.0004, 100.0005, 110., 200., 300., 400.0176], dtype='float')
-    intensities = numpy.array([1, 2, 3, 4, 5, 6, 7], dtype='float')
-    peak_comments = ["m/z 100.0003", None, "m/z 100.0005", "m/z 110.", "m/z 200.", "m/z 300.", "m/z 400.0176"]
-    peak_comments = {mz[i]: peak_comments[i] for i in range(len(mz))}
-    spectrum = Spectrum(mz=mz, intensities=intensities,
-                        metadata={"peak_comments": peak_comments})
-
-    spectrum.peaks = Spikes(mz=numpy.array([100.0004, 110., 400.018], dtype='float'),
-                            intensities=numpy.array([5, 4, 7], dtype='float'))
-
-    assert spectrum.peak_comments == {100.0004: "m/z 100.0003; m/z 100.0005", 110.: "m/z 110.", 400.018: "m/z 400.0176"}
